backend/ml-service/models/image_analyzer.py
"""
Wound Image Analyzer
Rule-based image analysis to classify wound severity
This is a placeholder for actual CNN model (e.g., transfer learning with ResNet)
"""

from PIL import Image
import io

def analyze_wound_image(image_bytes):
    """
    Analyze wound image and return severity classification
    
    Args:
        image_bytes (bytes): Image file bytes
    
    Returns:
        dict: {severity: str, score: int, confidence: float}
    """
    try:
        # Open image to validate it's a real image
        img = Image.open(io.BytesIO(image_bytes))
        
        # Get image properties
        width, height = img.size
        format_type = img.format
        
        # Rule-based classification (placeholder for actual CNN)
        # In real implementation, you'd use a trained CNN model here
        
        # For now, use image properties as a proxy
        # Larger images or certain formats might indicate more detailed photos
        pixel_count = width * height
        
        # Simple rule-based scoring
        if pixel_count > 2000000:  # Large detailed image
            severity = "severe"
            score = 85
        elif pixel_count > 1000000:  # Medium image
            severity = "moderate"
            score = 60
        else:  # Small image
            severity = "mild"
            score = 30
        
        # Add some randomness to simulate model variation
        import random
        score += random.randint(-10, 10)
        score = max(0, min(100, score))  # Clamp to 0-100
        
        return {
            "severity": severity,
            "score": score,
            "confidence": 0.75,
            "image_info": {
                "width": width,
                "height": height,
                "format": format_type
            }
        }
        
    except Exception as e:
        print(f"Error analyzing image: {str(e)}")
        # Return default mild severity on error
        return {
            "severity": "mild",
            "score": 25,
            "confidence": 0.5,
            "error": str(e)
        }
